+++ packages/core/src/cache/cache.ts
@@ -753,6 +753,7 @@ class Cache implements ICache {
     this.incrementImageCacheSize(-cachedImage.sizeInBytes);
 
     const eventDetails = {
+      image: cachedImage,
       imageId,
     };