+++ packages/tools/src/eventListeners/annotations/index.ts
@@ -1,4 +1,9 @@
+import annotationCompletedListener from './annotationCompletedListener';
 import annotationSelectionListener from './annotationSelectionListener';
 import annotationModifiedListener from './annotationModifiedListener';
 
-export { annotationSelectionListener, annotationModifiedListener };
+export {
+  annotationCompletedListener,
+  annotationSelectionListener,
+  annotationModifiedListener,
+};