+++ packages/adapters/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.50.3](https://github.com/dcmjs-org/dcmjs/compare/v1.50.2...v1.50.3) (2024-01-26)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.50.2](https://github.com/dcmjs-org/dcmjs/compare/v1.50.1...v1.50.2) (2024-01-26)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.50.1](https://github.com/dcmjs-org/dcmjs/compare/v1.50.0...v1.50.1) (2024-01-26)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+# [1.50.0](https://github.com/dcmjs-org/dcmjs/compare/v1.49.2...v1.50.0) (2024-01-25)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
 ## [1.49.2](https://github.com/dcmjs-org/dcmjs/compare/v1.49.1...v1.49.2) (2024-01-24)
 
 **Note:** Version bump only for package @cornerstonejs/adapters